requirements.txt
# Production
Django>=1.6.4
Pillow>=2.4.0
django-bootstrap3>=4.4.1
django-filebrowser>=3.5.6
django-grappelli>=2.5.3

# Develop
Werkzeug>=0.9.4
django-debug-toolbar>=1.2
django-extensions>=1.3.6
ipdb>=0.8

# Test
mixer>=4.0.7
pytest-django>=2.6.2
